Extract thread in to it's own component to be reused
