+++ nextjs/components/Pages/ThreadPage/index.tsx
@@ -1,0 +1,1 @@
+export { ThreadPage } from './ThreadPage';
