+++ nextjs/components/Thread/index.tsx
@@ -1,0 +1,1 @@
+export { Thread } from './Thread';
